Fix array initialization error
Set array to 0 in case `mpz_export` does not overwrite every field.

Example:
`mpz_export` exports a 31 byte value to a 32 byte array. This fix makes
sure, that the most segnificant byte is set to zero. A leftover value
other than 0 would result in a wrong value of the byte array which
caused the verification step to fail sometimes.